Multithread parts of the startup code
FindFMs.Find() now runs in parallel to the main window's initialization code. The two may or may not be close to overlapping in time taken depending on many factors, but nevertheless we at least cut the startup time by the smaller of the two run times. The improvement grows as the user's FM collection grows. On my setup, that's a savings of 60s warm. Nice.