ifdefs are now generated for InitializeComponent() implementation
Compiler's too dumb to realize it can optimize out a private method whose only callsite in inside a non-hit ifdef block, so I have to put another one right smack in the middle of the generated Designer.cs file that will constantly try to remove it. Awesome. But we shave like 28k off the exe like this, so we gotta.