Change delete FM menu icon
Now it's a trash can and is a more muted red, so that it more clearly matches what it does, and so the user can see it's dangerous-ish but isn't blinded by a bright red warning icon every time they simply open the FM menu. And now they won't be desensitized to the red exclamation mark icon either.