Remove early-out from SetFilter(), and remove Filtered bool
Our early-out was almost never being run because it required ShowUnsupported to be enabled, and that's meant to be disabled 99% of the time. Therefore, in practice we won't lose any time by simply always indirecting through the filtered index list, and we'll even save a little bit of time in the 99% case by not doing the early-out check.