Change all "C:\AngelLoader" to "C:\AngelLoader_dev_3053BA21"
To make it not overwrite anyone's install at "C:\AngelLoader". I personally just have the new folder hardlinked to the old, so no disruption for me.

A better idea: Maybe have a separate batch file that's local-only (ie. git-ignored) that's only run if it's found, and that copies the stuff.