RTF simplification with hooks
Removed the \cf0 inserts and the explicit first \colortbl entry set (when it would have been ";" / default), because with the hook we can now set its foreground to whatever we want, and independently of the other controls' hooked colors too. As a bonus we now finally have the right text highlight color for RTF too.