7z scans now cache readmes as part of the scan
Doing so is basically free because we already have the readmes extracted, and it prevents the situation where a long 7z scan extract takes place, and then it goes and caches the readmes separately which necessitates another long 7z extract. Saves a large amount of time on T2XHD.7z for example.