Config.Columns to array
Slight behavior change: Duplicate column entries in the config in file will now overwrite existing ones, rather than being ignored. But since this is clearly an unintended case that would only happen if someone manually edited their file for some reason, we're not going to care. The reason we're doing it is to keep the code simple and not have to check for already-filled columns.