FMData and config to custom fast delegate*<> dicts
FMData is no longer using the perfect hash, because while it's fast, it's also a pain in the ass to generate. We now use a dictionary of delegate*<> wrappers with a custom comparer that compares up the first '=' without having to take the allocation of a string split. This seems to be about as fast or at least fast enough to be basically instant even with ~1700 FMs, so well worth whatever tradeoff we're probably not even making.