Add notes about an issue that must be avoided
Having MainForm._displayedFM reference a stale FM object from the old list upon list rebuild, is something that could happen but doesn't because of us force-displaying the FM in all the right places through sheer luck. We should figure out a way to make this happen automatically when needed, to avoid future mistakes.