Fix stupid DGV cell double-click bug
If you double-click across rows (like click once on one row and then quickly move the mouse to the next row and click again within the double-click window), it would count as a double-click on the row selected by the second click, even though double-clicks should only count if you stay on the same row. It's been in there forever, seems like an internal bug, but only on some days can I get it to happen. It happens all the time or doesn't, depending on the phase of the moon that day. Whatever... force-fixed it with a stupid hack as usual.