Use FontEntries Top field instead of Count
And we can get rid of Count, and therefore the line incrementing it too. Now we check Top != null rather than Count > 0, so same number of external checks, one less internal increment, one less internal field, and replace a Count = 0 with a Top = null in Clear(), so equal on that.